feat(calculator): decompose calculator page orchestrator
